Add implementation of AESM client for SGX
This commit allows for using the AESM client from inside the enclave

Add building of aesm for SGX target on linux in CI

Add tests to build aesm-client during CI for all platforms

Added a common aesm_proto mod for sgx and unix
